iostream.[ch]: add fetch into user-specified buffer
One big problem with the io_fetch functions is that data will
be overwritten as I/O progresses. This turned out to be a problem
when reading large UV variables since the buffered information
describing the visdata record was lost while reading variable values.
Fix this by doing things the right way and making it possible to
read directly into a user-supplied buffer.