uvdecode.c: make it possible to turn off output at compile-time
For performance testing it is useful to have uvdecode not print
anything out. I've made it possible to disable this at compile
time. This should of course be a run-time option, but I'm lazy
and don't want to deal with argv handling in a quickie C program
right at the moment.