dataset.[ch]: fix multiple bugs in ds_list_items
Firstly, the only way to distinguish between an empty dataset
and an error listing the dataset was to check the GError parameter,
which is not desirable behavior.

Secondly, the code was reporting "header" as a dataset item.

Thirdly, the code to bail upon an item present as both small and
large form didn't actually bail.

Oops.